chore(clean-types): splits out the job and hook execution types, ahead of building a better public api for type exports
