+++ runem/job_execute.py
@@ -12,14 +12,11 @@ from runem.informative_dict import ReadOnlyInformativeDict
 from runem.job import Job
 from runem.job_wrapper import get_job_wrapper
 from runem.log import error, log
-from runem.types import (
-    FilePathList,
-    FilePathListLookup,
+from runem.types import FilePathList, FilePathListLookup, JobConfig, JobTags
+from runem.types_wip.types_jobs import (
     HookSpecificKwargs,
-    JobConfig,
     JobFunction,
     JobReturn,
-    JobTags,
     JobTiming,
     TimingEntries,
     TimingEntry,
