+++ runem/report.py
@@ -4,14 +4,13 @@ from collections import defaultdict
 from datetime import timedelta
 
 from runem.log import log
-from runem.types import (
+from runem.types import OrderedPhases, PhaseName
+from runem.types_wip.types_jobs import (
     JobReturn,
     JobRunMetadatasByPhase,
     JobRunReportByPhase,
     JobRunTimesByPhase,
     JobTiming,
-    OrderedPhases,
-    PhaseName,
     ReportUrlInfo,
     ReportUrls,
     TimingEntries,
