+++ scripts/test_hooks/py.py
@@ -6,7 +6,13 @@ from typing_extensions import Unpack
 
 from runem.log import log
 from runem.run_command import RunCommandUnhandledError, run_command
-from runem.types import FilePathList, JobKwargs, JobName, JobReturnData, OptionsWritable
+from runem.types_public_api import (
+    FilePathList,
+    JobKwargs,
+    JobName,
+    JobReturnData,
+    OptionsWritable,
+)
 
 
 def _job_py_code_reformat(
