+++ tests/test_types/test_public_api.py
@@ -1,0 +1,5 @@
+"""A really noddy file that documents which types are in the public API."""
+
+from runem.types_public_api import (  # noqa: F401 # pylint: disable=unused-import
+    FilePathList,
+)
